functions/requirements.txt
firebase-functions
firebase-admin
google-auth
google-api-python-client
requests
functions-framework
Flask